fix: fix building errors
