+++ src/lib/auth.ts
@@ -1,5 +1,5 @@
+import { authOptions } from "@/config/next-auth";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 export async function getSession() {
   return await getServerSession(authOptions);
